Add convert tool
This is the new equivalent to use-local for download. download's
use-local has fallen into disuse (it doesn't even work right now) and
will diverge more with further changes.

This is not really intended to be used for current data as the EDN
files in this repo is now considered the source of truth. However,
this will be extended to import translations from json files which is
intended to be kept in sync between jinteki and netrunner-cards-json.